Fix scoping and variable issues
